Rename ext to extension.
