+++ simphony/cuds/simulation.py
@@ -1,6 +1,6 @@
 """Contains simulation controllers."""
 from . import CUDS
-from ..ext import create_wrapper
+from ..extension import create_wrapper
 
 
 class Simulation(object):
